harvester/requirements.txt
antiorm==1.2.1
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorama==0.4.1
config==0.4.2
connexion==2.4.0
coverage==4.5.4
DateTime==4.3
db==0.1.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
gp==1.0
gunicorn==20.0.3
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
logger==1.4
lxml==4.4.2
MarkupSafe==1.1.1
marshmallow==3.2.2
marshmallow-sqlalchemy==0.19.0
mccabe==0.6.1
more-itertools==7.2.0
openapi-spec-validator==0.2.8
packaging==19.2
pipenv==2018.11.26
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pyrsistent==0.15.6
pytest==5.2.1
pytest-cov==2.8.1
pytz==2019.3
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
simplejson==3.17.0
six==1.12.0
soupsieve==1.9.5
SQLAlchemy==1.3.11
urllib3==1.25.7
virtualenv==16.7.5
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.16.0
wget==3.2
yahoo-finance==1.4.0
ystockquote==0.2.5
zipp==0.6.0
zope.interface==4.7.1
